+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 - Add new cop `RSpec/LeakyLocalVariable`. ([@lovro-bikic])
+- Fix detection of nameless doubles with methods in `RSpec/VerifiedDoubles`. ([@ushi-as])
 
 ## 3.7.0 (2025-09-01)
 
@@ -1079,6 +1080,7 @@ Compatibility release so users can upgrade RuboCop to 0.51.0. No new features.
 [@tmaier]: https://github.com/tmaier
 [@topalovic]: https://github.com/topalovic
 [@twalpole]: https://github.com/twalpole
+[@ushi-as]: https://github.com/ushi-as
 [@vzvu3k6k]: https://github.com/vzvu3k6k
 [@walf443]: https://github.com/walf443
 [@yasu551]: https://github.com/yasu551